requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
distro==1.9.0
filetype==1.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.1
lxml==5.3.0
openai==1.54.4
pydantic==2.9.2
pydantic_core==2.23.4
pyotp==2.9.0
PyYAML==6.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
tqdm==4.67.0
twikit==2.1.3
typing_extensions==4.12.2
